+++ src/index.js
@@ -1,12 +1,12 @@
 import 'babel-polyfill';
 
-import DataFrame from './dataframe.js';
-import Row from './row.js';
-import * as Errors from './errors.js';
-import Stat from './modules/stat.js';
-import Matrix from './modules/matrix.js';
-import SQL from './modules/sql.js';
-import Benchmark from './extras/benchmark.js';
+import DataFrame from './dataframe';
+import Row from './row';
+import * as Errors from './errors';
+import Stat from './modules/stat';
+import Matrix from './modules/matrix';
+import SQL from './modules/sql';
+import Benchmark from './extras/benchmark';
 
 DataFrame.setDefaultModules(Stat, Matrix, SQL);
 DataFrame.sql = SQL;